gilded_rose.py: make quality up/down logic more similar
In preparation for extracting a simple method, interchange a few
`if` statements so that when `item.quality` is modified it's always
with the format

    if item.quality is in bounds:
    	item.quality = item.quality +/- 1

Also, remove a pointless subtraction that always evaluated to zero,
and replace with the literal constant zero.

Verified by re-running unit test and text based regression.