characterization tests: backstage passes
The quality adjustments for backstage passes have a number of
boundary conditions:

- initial quality value
- sell_in days remaining

Add tests for a variety of these.

I normally don't like tests like this one that do many individual tests
internally, but there are enough boundary cases that I think this is the
right approach. To make the failure messages more useful, I've added an
extra output string on the asserts indicating which subtest is failing.