gilded_rose: extract logic for aged brie
Modify `update_quality` so that the logic for aged brie is entirely
contained in one block at the to of the for loop and the rest of the for
loop is skipped.

The remainder of the loop is unchanged, but all conditions of the form

    item.name != "Aged Brie"

will be true, since that case now never reaches here. The next commit
will remove those unnecessary comparisons.