gilded_rose.py: initial support for conjured items
- test_gilded_rose.py: first tests for basic conjured behaviour
- gilded_rose.py: implementation that passes the tests

I meant to make a few more commits before this one to demonstrate
writing a failing test first, and then creating the code needed to pass
it, and *just* that code. A quick summary of ConjuredItemTests

- test_quality_drops_by_2 was written first, and fails before any
  other changes are made
- then, I modified the generic item code to check for a "Conjured"
  prefix, and multiply the decrement by 2 *just* for the initial call to
  decrement_quality. This is all that's needed to make the first test
  pass.
- add the second test, which forces multiplying the decrement by 2 in
  the sell_in < 0 case.

This commit is probably close to what I'd normally do, instead of a
commit for such tiny steps, but the key point is I literally went
through the steps described above, just without intervening commits.